move components to their respective directories
